% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aipw.ate.R
\name{aipw.ate}
\alias{aipw.ate}
\title{Combines IPW and IMP estimators to form the augmented IPW, AIPW}
\usage{
aipw.ate(y, treated, imp, ipw)
}
\arguments{
\item{y}{Observed response}

\item{treated}{Binary vetor indicating treatment}

\item{imp}{imp_output object from imp.ate()}

\item{ipw}{ipw_output object from ipw.ate()}
}
\value{
Average treatment effect (ATE) for the augmented IPW (AIPW)
}
\description{
Augmented IPW (AIPW) as in Ghosh, Ma, & De Luna (2020).
}
\examples{
# Using example data from package SDRcausal
library(SDRcausal)

# Import example data
x <- SDRcausal::covariates
y <- SDRcausal::outcomes
trt <- SDRcausal::treated
b1 <- SDRcausal::beta1_guess
b0 <- SDRcausal::beta0_guess
alp <- SDRcausal::alpha_guess

# Perform semiparametric imputation
imp <- SDRcausal::imp.ate(x, y, trt, b1, b0,
           explicit_bandwidth = TRUE, bwc_dim_red1 = 1, bwc_impute1 = 1,
           bwc_dim_red0 = 1, bwc_impute0 = 1)

# Perform semiparametric inverse probability weighting
ipw <- SDRcausal::ipw.ate(x, y, trt, alp, bwc_dim_red = 8,
           bwc_prop_score = 8)

# Calculate the Augmented IPW (AIPW)
aipw <- SDRcausal::aipw.ate(y, trt, imp, ipw)

}
\references{
Ghosh, T., Ma, Y., & De Luna, X. (2020). Sufficient dimension
reduction for feasible and robust estimation of average causal effect.
Statistica Sinica, accepted.
}
