% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aipw2.ate.R
\name{aipw2.ate}
\alias{aipw2.ate}
\title{Improved Augmented IPW (AIPW2)}
\usage{
aipw2.ate(y, treated, imp, ipw)
}
\arguments{
\item{y}{Observed response}

\item{treated}{A binary vetor indicating treatment}

\item{imp}{imp output object from \code{imp.ate}}

\item{ipw}{ipw output object from \code{ipw.ate}}
}
\value{
The improved augmented IPW estimation (AIPW2) of the average treatment effect (ATE).
}
\description{
Combines IPW and IMP estimators to form the improved augmented
             IPW, AIPW2 as in Ghosh, Ma, & De Luna (2020).
}
\examples{
# Using example data from package SDRcausal
library(SDRcausal)

# Import example data
x <- SDRcausal::covariates
y <- SDRcausal::outcomes
trt <- SDRcausal::treated
b1 <- SDRcausal::beta1_guess
b0 <- SDRcausal::beta0_guess
alp <- SDRcausal::alpha_guess

# Perform semiparametric imputation
imp <- SDRcausal::imp.ate(x, y, trt, b1, b0,
           explicit_bandwidth = TRUE, bwc_dim_red1 = 1, bwc_impute1 = 1,
           bwc_dim_red0 = 1, bwc_impute0 = 1)

# Perform semiparametric inverse probability weighting
ipw <- SDRcausal::ipw.ate(x, y, trt, alp, bwc_dim_red = 10,
           bwc_prop_score = 18)

# Calculate the Improved Augmented IPW (AIPW2)
aipw2 <- SDRcausal::aipw2.ate(y, trt, imp, ipw)

}
\references{
Ghosh, T., Ma, Y., & De Luna, X. (2020). Sufficient dimension
reduction for feasible and robust estimation of average causal effect.
Statistica Sinica, accepted.
}
