% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpha_guess}
\alias{alpha_guess}
\title{Study 1 dataset.}
\format{
A vector with 6 terms:
}
\usage{
alpha_guess
}
\description{
A dataset containing the initial guess for alpha.
}
\keyword{datasets}
