% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_utilities.R
\name{b10_fun}
\alias{b10_fun}
\title{Calculates B1/0}
\usage{
b10_fun(x_lower, treated, dm, beta, kernel, bandwidth, gauss_cutoff)
}
\arguments{
\item{x_lower}{Projection of covariate matrix on CMS}

\item{treated}{Binary vector indicating treatment.}

\item{dm}{Derivative of imputed values}

\item{beta}{CMS}

\item{kernel}{Specifies which kernel function to be used}

\item{bandwidth}{Specifies if bandwidth_scale will be used as the}

\item{gauss_cutoff}{Cutoff value for Gaussian kernel}
}
\value{
B1/0 matrix
}
\description{
Calculates Eq 2.8 or 2.10 in Ghosh, Ma, & De Luna (2020).
}
