% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beta1_guess}
\alias{beta1_guess}
\title{Study 1 dataset.}
\format{
A vector with 6 terms:
}
\usage{
beta1_guess
}
\description{
A dataset containing the initial guess for beta 1.
}
\keyword{datasets}
