% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covariates}
\alias{covariates}
\title{Study 1 dataset.}
\format{
A data frame with 1000 rows and 6 variables:
\describe{
  \item{X1}{first covariate}
  \item{X2}{second covariate}
  ...
}
}
\usage{
covariates
}
\description{
A dataset containing the covariets.
}
\keyword{datasets}
