% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp.var.R
\name{imp.var}
\alias{imp.var}
\title{Estimates IMP variance}
\usage{
imp.var(
  x,
  y,
  treated,
  imp,
  ipw,
  bandwidth_scale1 = imp$bw1,
  bandwidth_scale0 = imp$bw0,
  kernel = "EPAN",
  explicit_bandwidth = TRUE,
  gauss_cutoff = 0.001
)
}
\arguments{
\item{x}{Covariate matrix}

\item{y}{Response vector}

\item{treated}{A binary vetor indicating treatment}

\item{imp}{imp output object from \code{imp.ate}}

\item{ipw}{ipw output object from \code{ipw.ate}}

\item{bandwidth_scale1}{Scaling of the calculated bandwidth, or in case of \code{explicit_bandwidth = TRUE}, the actual bandwidth for the estimation of \eqn{E(\cdot|\beta_1^T X)}. The default value is \code{imp$bw1}. If this default value is used, one should use the default value \code{TRUE} for \code{explicit_bandwidth}.}

\item{bandwidth_scale0}{Scaling of the calculated bandwidth, or in case of \code{explicit_bandwidth = TRUE}, the actual bandwidth for the estimation of \eqn{E(\cdot|\beta_0^T X)}. The default value is \code{imp$bw0}. If this default value is used, one should use the default value \code{TRUE} for \code{explicit_bandwidth}.}

\item{kernel}{Specifies which kernel function to be used. The default is \code{"EPAN"}.}

\item{explicit_bandwidth}{Specifies if bandwidth_scale will be used as the bandwidth or if it will be calculated as \code{bandwidth_scale} * sd(\eqn{\beta_t^T x}) * \eqn{n^{(1/5)}}. The default value is \code{TRUE}.}

\item{gauss_cutoff}{The cutoff value for Gaussian kernel. The default value is \code{1e-3}.}
}
\value{
Variance of IMP estimator
}
\description{
Variance of the IMP as in Ghosh, Ma, & De Luna (2020).
}
\examples{
# Using example data from package SDRcausal
library(SDRcausal)

# Import example data
x <- SDRcausal::covariates
y <- SDRcausal::outcomes
trt <- SDRcausal::treated
b1 <- SDRcausal::beta1_guess
b0 <- SDRcausal::beta0_guess
alp <- SDRcausal::alpha_guess

# Perform semiparametric imputation
imp <- SDRcausal::imp.ate(x, y, trt, b1, b0,
           explicit_bandwidth = TRUE, bwc_dim_red1 = 1, bwc_impute1 = 1,
           bwc_dim_red0 = 1, bwc_impute0 = 1)

# Perform semiparametric inverse probability weighting
ipw <- SDRcausal::ipw.ate(x, y, trt, alp, bwc_dim_red = 10,
           bwc_prop_score = 18)

# Calculate the variance of the IMP estimator.
var <- SDRcausal::imp.var(x, y, trt, imp, ipw,
           bandwidth_scale1 = imp$bw1, bandwidth_scale0 = imp$bw0)

}
\references{
Ghosh, T., Ma, Y., & De Luna, X. (2020). Sufficient dimension
reduction for feasible and robust estimation of average causal effect.
Statistica Sinica, accepted.
}
