% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improved_augmented_ipw.R
\name{improved_augmented_ipw}
\alias{improved_augmented_ipw}
\title{Improved Augmented IPW (IAIPW)}
\usage{
improved_augmented_ipw(y, treated, imp, ipw)
}
\arguments{
\item{y}{Observed response}

\item{treated}{Binary vetor indicating treatment}

\item{imp}{imp_output object from semipar_imputation()}

\item{ipw}{ipw_output object from semipar_ipw()}
}
\value{
Average treatment effect (ATE) for the improved augmented IPW
        (IAIPW)
}
\description{
Combines IPW and IMP estimators to form the improved augmented
             IPW, IAIPW as in Ghosh, Ma, & De Luna (2020).
}
\examples{
# Using example data from package SDRcausal
library(SDRcausal)

# Import example data
x <- SDRcausal::covariates
y <- SDRcausal::outcomes
trt <- SDRcausal::treated
b1 <- SDRcausal::beta1_guess
b0 <- SDRcausal::beta0_guess
alp <- SDRcausal::alpha_guess

# Perform semiparametric imputation
imp <- SDRcausal::semipar_imputation(x, y, trt, b1, b0,
           explicit_bandwidth = TRUE, bwc_dim_red1 = 1, bwc_impute1 = 1,
           bwc_dim_red0 = 1, bwc_impute0 = 1)

# Perform semiparametric inverse probability weighting
ipw <- SDRcausal::semipar_ipw(x, y, trt, alp, bwc_dim_red = 10,
           bwc_prop_score = 18)

# Calculate the variance of the Augmented IPW (AIPW)
iaipw <- SDRcausal::improved_augmented_ipw(y, trt, imp, ipw)

}
\references{
Ghosh, T., Ma, Y., & De Luna, X. (2020). Sufficient dimension
reduction for feasible and robust estimation of average causal effect.
Statistica Sinica, accepted.
}
