% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Estimates imputed values based on CMS}
\usage{
impute(
  x,
  y,
  treated,
  beta_hat,
  kernel = "EPAN",
  explicit_bandwidth = FALSE,
  bandwidth_scale = 1,
  gauss_cutoff = 0.001,
  to_extrapolate = TRUE,
  to_truncate = TRUE,
  extrapolation_basis = as.integer(5),
  verbose = FALSE
)
}
\arguments{
\item{x}{Covariate matrix}

\item{y}{Response vector}

\item{treated}{Binary vetor indicating treatment}

\item{beta_hat}{Locally efficient CMS}

\item{kernel}{Specifies which kernel function to be used}

\item{explicit_bandwidth}{Specifies if bandwidth_scale will be used as the bandwidth or
if it will be calculated as bw = bandwidth_scale
* sd(x * beta) * n^(1/3)}

\item{bandwidth_scale}{Kernel bandwidth}

\item{gauss_cutoff}{Cutoff value for Gaussian kernel}

\item{to_extrapolate}{Specifies wheter to extrapolate or not}

\item{to_truncate}{Specifies wheter to extrapolate or not}

\item{extrapolation_basis}{Number of data point to base extrapolation on.}

\item{verbose}{Specifies if the program should print output while
running}
}
\value{
A list containing the reduced space xb, the imputed values and their
        derivatives.
}
\description{
Performs semiparametric imputation based on the CMS calculated
             by imp_dim_red, as in Ghosh, Ma, & De Luna (2020).
}
\examples{
# Using example data from package SDRcausal
library(SDRcausal)

# Import example data
x <- SDRcausal::covariates
y <- SDRcausal::outcomes
trt <- SDRcausal::treated
b1 <- SDRcausal::beta1_guess
b0 <- SDRcausal::beta0_guess

# Using example data from package SDRcausal
library(SDRcausal)

# Import example data
x <- SDRcausal::covariates
y <- SDRcausal::outcomes
trt1 <- SDRcausal::treated
n <- as.integer(dim(x)[1])
trt0 <- as.integer(rep(1, times = n) - trt1)
b1 <- SDRcausal::beta1_guess
b0 <- SDRcausal::beta0_guess

# Perform semiparametric dimension reduction for treated
cms1 <- SDRcausal::imp_dim_red(x, y, trt1, b1,
           explicit_bandwidth = 1, bandwidth_scale = 1)

# Perform semiparametric dimension reduction for untreated
cms0 <- SDRcausal::imp_dim_red(x, y, trt0, b0,
           explicit_bandwidth = 1, bandwidth_scale = 1)

# Perform semiparametric imputation for treated
m1 <- SDRcausal::impute(x, y, trt1, cms1$fb,
         explicit_bandwidth = 1, bandwidth_scale = cms1$bw)

# Perform semiparametric imputation for untreated
m0 <- SDRcausal::impute(x, y, trt0, cms0$fb,
         explicit_bandwidth = 1, bandwidth_scale = cms0$bw)

}
\references{
Ghosh, T., Ma, Y., & De Luna, X. (2020). Sufficient dimension
reduction for feasible and robust estimation of average causal effect.
Statistica Sinica, accepted.
}
