% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipw.ate.R
\name{ipw.ate}
\alias{ipw.ate}
\title{Estimates average treatment effect through IPW}
\usage{
ipw.ate(
  x,
  y,
  treated,
  alpha_initial,
  solver = "optim",
  kernel = "EPAN",
  explicit_bandwidth = FALSE,
  recalc_bandwidth = TRUE,
  bwc_dim_red = 1,
  bwc_prop_score = 10,
  gauss_cutoff = 0.001,
  penalty = 10,
  n_before_pen = 1,
  n_threads = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Covariate matrix}

\item{y}{Response vector}

\item{treated}{A binary vector indicating treatment.}

\item{alpha_initial}{Initial guess for \eqn{\alpha}}

\item{solver}{Specifies which solver is to be used. Current options are \code{optim} and \code{cobyla} (from \code{nloptr} package). The diffault value is \code{"optim"}.}

\item{kernel}{Specifies which kernel function is to be used, current options are: \code{"EPAN"}, \code{"QUARTIC"}, and \code{"GAUSSIAN"}. The default value is \code{"EPAN"}.}

\item{explicit_bandwidth}{Specifies if \code{bandwidth_scale} will be used as the bandwidth or if it will be calculated as \code{bandwidth_scale} * sd(\eqn{\alpha^T x}) * \eqn{n^{(1/5)}}. The default value is \code{FALSE}.}

\item{recalc_bandwidth}{Specifies whether the bandwidth should be recalculated after the estimations of \eqn{\alpha}. If \code{explicit_bandwidth} is \code{TRUE}, \code{recalc_bandwidth} is not used, but if \code{explicit_bandwidth} is \code{FALSE}, then if \code{recalc_bandwidth} is \code{TRUE}, bandwidths are recalculated at the beginning of the second step based on \code{bwc_prop_score}. If \code{recalc_bandwidth} is \code{FALSE}, the first step bandwidth is used. The default value is \code{TRUE}.}

\item{bwc_dim_red}{Scaling of calculated bandwidth, or if \code{explicit_bandwidth = TRUE} used as the bandwidth. It is used in the dimension reduction step for \eqn{\alpha^T x}. The default value is \code{1}.}

\item{bwc_prop_score}{Scaling of calculated bandwidth, or if \code{explicit_bandwidth = TRUE} used as the bandwidth. It is used for the estimation of the propensity score. The default value is \code{10}.}

\item{gauss_cutoff}{The cutoff value for Gaussian kernel. The default value is \code{1e-3}.}

\item{penalty}{Penalty for the optimizer if a probability is outside (0, 1) during the estimation of \eqn{\alpha} phase. Added to the function value in solver as \code{penalty}^(n - \code{n_before_pen}), where n is the number of probabilities outside (0, 1). The default value is \code{10}.}

\item{n_before_pen}{The number of probabilities outside the range (0, 1) to accept during the estimation of \eqn{\alpha} phase. The default value is \code{1}.}

\item{n_threads}{Sets the number of threads for parallel computing. Set to 1 serial. If \code{n_threads} exceeds the maximum number of threads, sets \code{n_threads} to max_threads - 1. To use max_threads, set to \code{n_threads} to max_threads of system. The default value is \code{1}.}

\item{verbose}{Specifies if the program should print output while running. The default value is \code{FALSE}.}

\item{...}{Additional parameters passed to \code{optim} or \code{cobyla}.}
}
\value{
A list containing the average treatment effect (ate), the propensity
        score (pr), the final alpha (fa), and the output from optim (op).
}
\description{
Semiparametric estimation of the average treatment effect based
             on the IPW method described in Ghosh, Ma, & De Luna (2020).
}
\examples{
# Using example data from package SDRcausal
library(SDRcausal)

# Import example data
x <- SDRcausal::covariates
y <- SDRcausal::outcomes
trt <- SDRcausal::treated
alp <- SDRcausal::alpha_guess

# Perform semiparametric inverse probability weighting
ipw <- SDRcausal::ipw.ate(x, y, trt, alp)

}
\references{
Ghosh, T., Ma, Y., & De Luna, X. (2020). Sufficient dimension
reduction for feasible and robust estimation of average causal effect.
Statistica Sinica, accepted.
}
\seealso{
[stats::optim]
}
