% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipw_dim_red.R
\name{ipw_dim_red}
\alias{ipw_dim_red}
\title{Estimates the Central Mean Space (CMS)}
\usage{
ipw_dim_red(
  x,
  treated,
  alpha_initial,
  solver = "optim",
  kernel = "EPAN",
  explicit_bandwidth = FALSE,
  bandwidth_scale = 1,
  gauss_cutoff = 0.001,
  penalty = 10,
  n_before_pen = 5,
  root_tol = 0.001,
  n_threads = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Covariate matrix}

\item{treated}{Binary vetor indicating treatment}

\item{alpha_initial}{Initial guess of CMS}

\item{solver}{Specifies which solver to be used. Current options
optim and cobyla (from nloptr package).}

\item{kernel}{Specifies which kernel function to be used,
current options are: "EPAN", "QUARTIC", and
"GAUSSIAN".}

\item{explicit_bandwidth}{Specifies if bandwidth_scale will be used
explicitly as the bandwidth.}

\item{bandwidth_scale}{Scaling of the calculated bandwidth, or in case of
explicit_bandwidth = TRUE the bandwidth.}

\item{gauss_cutoff}{cutoff value for Gaussian kernel}

\item{penalty}{Penalty for the optimizer if a probability is
outside (0, 1). Added to the function value in
optim as: penalty^(n), where n is the number of
probabilities outside (0, 1).}

\item{n_before_pen}{Number of probabilities outside the range (0, 1)
to accept during dimension reduction.}

\item{root_tol}{Tolerance which makes the program warn if optim
stops at at a value higher than root_tol.}

\item{n_threads}{Sets number of threads for parallel run. Set to 0
serial. If n_threads exceeds maximum number of
threads, sets n_threads to max_threads - 1. To use
max_threads, set to n_threads to max_threads of
system.}

\item{verbose}{Specifies if the program should print output while
running.}

\item{...}{Additional parameters passed to solver.}
}
\value{
A list containing the final alpha, bandwwidth used, and the output
        of optim
}
\description{
Semiparametric estimation of the Central Mean Space (CMS) as in
             Ghosh, Ma, & De Luna (2020). To be used with SDRcausal::propensity_score().
}
\examples{
# Using example data from package SDRcausal
library(SDRcausal)

# Import example data
x <- SDRcausal::covariates
y <- SDRcausal::outcomes
trt <- SDRcausal::treated
alp <- SDRcausal::alpha_guess

# Perform semiparametric dimension reduction for treated
cms <- SDRcausal::ipw_dim_red(x, trt, alp,
           explicit_bandwidth = TRUE, bandwidth_scale = 8)

}
\references{
Ghosh, T., Ma, Y., & De Luna, X. (2020). Sufficient dimension
reduction for feasible and robust estimation of average causal effect.
Statistica Sinica, accepted.
}
\seealso{
[stats::optim]
}
