% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.imp}
\alias{plot.imp}
\title{Plots imputation output}
\usage{
\method{plot}{imp}(x, ..., covariates, y, treated)
}
\arguments{
\item{x}{imp_output object from imp.ate()}

\item{...}{Other parameters}

\item{covariates}{Covariate matrix}

\item{y}{Response vector}

\item{treated}{Binary vetor indicating treatment}
}
\value{
A list of ggplot plots of observed and imputed values (pl_imp),
        imputed treated values vs CMS (pl_m1), and imputed untreated values
        vs CMS (pl_m0).
}
\description{
Plot function for visualisation of imputation output from
             imp.ate. Note: The function requires ggplot2.
}
\examples{
# Using example data from package SDRcausal
library(SDRcausal)

# Import example data
covariates <- SDRcausal::covariates
y <- SDRcausal::outcomes
trt <- SDRcausal::treated
b1 <- SDRcausal::beta1_guess
b0 <- SDRcausal::beta0_guess
alp <- SDRcausal::alpha_guess

# Perform semiparametric imputation
imp <- SDRcausal::imp.ate(covariates, y, trt, b1, b0,
           explicit_bandwidth = TRUE, bwc_dim_red1 = 1, bwc_impute1 = 1,
           bwc_dim_red0 = 1, bwc_impute0 = 1)

# Plotting
plots <- plot(imp , covariates = covariates, y=y, treated = trt)

}
