% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.imp}
\alias{plot.imp}
\alias{plot.ipw}
\title{Plots imputation output}
\usage{
\method{plot}{imp}(x, y, treated, imp)

\method{plot}{ipw}(treated, ipw)
}
\arguments{
\item{x}{Covariate matrix}

\item{y}{Response vector}

\item{treated}{Binary vetor indicating treatment}

\item{imp}{imp_output object from imp.ate()}

\item{ipw}{ipw_output object from ipw.ate()}
}
\value{
A list of ggplot plots of observed and imputed values (pl_imp),
        imputed treated values vs CMS (pl_m1), and imputed untreated values
        vs CMS (pl_m0).

ggplot plot of the propensity score vs CMS.
}
\description{
Plot function for visualisation of imputation output from
             imp.ate. Note: The function requires ggplot2.

Plot function for visualisation of IPW output from ipw.ate.
             Note: The function requires ggplot2.
}
\examples{
# Using example data from package SDRcausal
library(SDRcausal)

# Import example data
x <- SDRcausal::covariates
y <- SDRcausal::outcomes
trt <- SDRcausal::treated
b1 <- SDRcausal::beta1_guess
b0 <- SDRcausal::beta0_guess
alp <- SDRcausal::alpha_guess

# Perform semiparametric imputation
imp <- SDRcausal::imp.ate(x, y, trt, b1, b0,
           explicit_bandwidth = TRUE, bwc_dim_red1 = 1, bwc_impute1 = 1,
           bwc_dim_red0 = 1, bwc_impute0 = 1)

# Plotting
plots <- plot(x, y, trt, imp)

# Using example data from package SDRcausal
library(SDRcausal)

# Import example data
x <- SDRcausal::covariates
y <- SDRcausal::outcomes
trt <- SDRcausal::treated
b1 <- SDRcausal::beta1_guess
b0 <- SDRcausal::beta0_guess
alp <- SDRcausal::alpha_guess

# Perform semiparametric imputation
ipw <- SDRcausal::ipw.ate(x, y, trt, alp, bwc_dim_red = 8,
           bwc_prop_score = 8)

# Plotting
plots <- plot(trt, ipw)

}
