% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_ipw}
\alias{plot_ipw}
\title{Plots IPW output}
\usage{
plot_ipw(treated, ipw)
}
\arguments{
\item{treated}{Binary vetor indicating treatment}

\item{ipw}{ipw_output object from ipw.ate()}
}
\value{
ggplot plot of the propensity score vs CMS.
}
\description{
Plot function for visualisation of IPW output from ipw.ate.
             Note: The function requires ggplot2.
}
\examples{
# Using example data from package SDRcausal
library(SDRcausal)

# Import example data
x <- SDRcausal::covariates
y <- SDRcausal::outcomes
trt <- SDRcausal::treated
b1 <- SDRcausal::beta1_guess
b0 <- SDRcausal::beta0_guess
alp <- SDRcausal::alpha_guess

# Perform semiparametric imputation
ipw <- SDRcausal::ipw.ate(x, y, trt, alp, bwc_dim_red = 8,
           bwc_prop_score = 8)

# Plotting
plots <- plot_ipw(trt, ipw)

}
