% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity_score.R
\name{propensity_score}
\alias{propensity_score}
\title{Estimates propensity score}
\usage{
propensity_score(
  x,
  treated,
  alpha_hat,
  kernel = "EPAN",
  explicit_bandwidth = FALSE,
  bandwidth_scale = 1,
  verbose = FALSE
)
}
\arguments{
\item{x}{Covariate matrix}

\item{treated}{Binary vetor indicating treatment}

\item{alpha_hat}{Locally efficient CMS}

\item{kernel}{Kernel specification}

\item{explicit_bandwidth}{Specifies if bandwidth_scale will be used as the
bandwidth or if it will be calculated as bw =
bandwidth_scale * sd(x * beta) * n^(1/3).}

\item{bandwidth_scale}{Scaling of calculated bandwidth, or if
explicit_bandwidth = TRUE used as the banddwidth.}

\item{verbose}{Specifies if the program should print output while
running.}
}
\value{
A list containing the estimated propensity scores values and their
        derivatives, and the bandwidth used.
}
\description{
Semiparametric estimation of the propensity score as in Ghosh,
             Ma, & De Luna (2020). To be used with SDRcausal::ipw_dim_red().
}
\examples{
# Using example data from package SDRcausal
library(SDRcausal)

# Import example data
x <- SDRcausal::covariates
y <- SDRcausal::outcomes
trt <- SDRcausal::treated
alp <- SDRcausal::alpha_guess

# Perform semiparametric dimension reduction
cms <- SDRcausal::ipw_dim_red(x, trt, alp,
           explicit_bandwidth = TRUE, bandwidth_scale = 8)

# Estimate propensity score
pr_score <- SDRcausal::propensity_score(x, trt, cms$fa,
               bandwidth_scale = 8)

}
\references{
Ghosh, T., Ma, Y., & De Luna, X. (2020). Sufficient dimension
reduction for feasible and robust estimation of average causal effect.
Statistica Sinica, accepted.
}
