% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipar_ipw.R
\name{semipar_ipw}
\alias{semipar_ipw}
\title{Estimates average treatment effect through IPW}
\usage{
semipar_ipw(
  x,
  y,
  treated,
  alpha_initial,
  kernel = "EPAN",
  explicit_bandwidth = FALSE,
  recalc_bandwidth = TRUE,
  bwc_dim_red = 1,
  bwc_prop_score = 1,
  gauss_cutoff = 0.001,
  penalty = 10,
  n_before_pen = 1,
  n_threads = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Covariate matrix}

\item{y}{Response vector}

\item{treated}{Binary vector indicating treatment.}

\item{alpha_initial}{Initial guess of beta for m1}

\item{kernel}{Specifies which kernel function to be used,
current options are: "EPAN", "QUARTIC", and
"GAUSSIAN".}

\item{explicit_bandwidth}{Specifies if bandwidth_scale will be used as the
bandwidth or if it will be calculated as bw =
bandwidth_scale * sd(x * beta) * n^(1/3).}

\item{recalc_bandwidth}{Specifies wheter the bandwidth should be
recalculated after the estimation of alpha
(ipw_dim_red)}

\item{bwc_dim_red}{Scaling of calculated bandwidth, or if
explicit_bandwidth = TRUE used as the banddwidth.
For dimension reduction (ipw_dim_red).}

\item{bwc_prop_score}{Scaling of calculated bandwidth, or if
explicit_bandwidth = TRUE used as the banddwidth.
Recalculated if explicit_bandwidth = FALSE and
recalc_bandwidth = TRUE. For propensity score.}

\item{gauss_cutoff}{cutoff value for Gaussian kernel}

\item{penalty}{Penalty for the optimizer if a probability is
outside (0, 1) during dimension reduction. Added
to the function value in solver as: penalty^(n -
n_before_pen), where n is the number of
probabilities outside (0, 1).}

\item{n_before_pen}{Number of probabilities outside the range (0, 1)
to accept during dimension reduction.}

\item{n_threads}{Sets number of threads for parallel run. Set to 0
serial. If n_threads exceeds maximum number of
threads, sets n_threads to max_threads - 1. To
use max_threads, set to n_threads to max_threads
of system.}

\item{verbose}{Specifies if the program should print output while
running.}

\item{...}{Additional parameters passed to optim.}
}
\value{
A list containing the average treatment effect (ate), the propensity
        score (pr), the final alpha (fa), and the output from optim (op).
}
\description{
Semiparametric estimation of the average treatment effect based
             on the IPW method described in Ghosh, Ma, & De Luna (2020).
}
\examples{
# Using example data from package SDRcausal
library(SDRcausal)

# Import example data
x <- SDRcausal::covariates
y <- SDRcausal::outcomes
trt <- SDRcausal::treated
alp <- SDRcausal::alpha_guess

# Perform semiparametric inverse probability weighting
ipw <- SDRcausal::semipar_ipw(x, y, trt, alp, bwc_dim_red = 8,
           bwc_prop_score = 8)

}
\references{
Ghosh, T., Ma, Y., & De Luna, X. (2020). Sufficient dimension
reduction for feasible and robust estimation of average causal effect.
Statistica Sinica, accepted.
}
\seealso{
[stats::optim]
}
