% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_utilities.R
\name{b_fun}
\alias{b_fun}
\title{Calculates B1/0}
\usage{
b_fun(x, d_eta, alpha_hat, h, kernel, bandwidth, verbose)
}
\arguments{
\item{x}{Projection of covariate matrix on CMS}

\item{d_eta}{Derivative of propensity score}

\item{alpha_hat}{Derivative of imputed values}

\item{h}{CMS}

\item{kernel}{Specifies which kernel function to be used}

\item{bandwidth}{Kernel bandwidth}

\item{verbose}{Specifies if the program should print output while running.}
}
\value{
B1/0 matrix
}
\description{
Calculates Eq 2.8 or 2.10 in Ghosh, Ma, & De Luna (2020).
}
