% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf.ate.R
\name{inf.ate}
\alias{inf.ate}
\title{Performs Estimations of Average Treatment Effect and Infrences}
\usage{
inf.ate(
  x,
  y,
  treated,
  beta_guess1,
  beta_guess0,
  imp.solver = "optim",
  imp.kernel = "EPAN",
  imp.explicit_bandwidth = FALSE,
  imp.recalc_bandwidth = TRUE,
  bwc_dim_red1 = 1,
  bwc_impute1 = 1.25,
  bwc_dim_red0 = 1,
  bwc_impute0 = 1.25,
  imp.gauss_cutoff = 0.001,
  imp.penalty = 10,
  imp.n_before_pen = 5,
  imp.to_extrapolate = TRUE,
  imp.to_truncate = TRUE,
  imp.extrapolation_basis = 5,
  alpha_initial,
  ipw.solver = "optim",
  ipw.kernel = "EPAN",
  ipw.explicit_bandwidth = FALSE,
  ipw.recalc_bandwidth = TRUE,
  bwc_dim_red = 1,
  bwc_prop_score = 10,
  ipw.gauss_cutoff = 0.001,
  ipw.penalty = 10,
  ipw.n_before_pen = 1,
  n_threads = 1,
  verbose = TRUE,
  imp.solver.options = NA,
  ipw.solver.options = NA
)
}
\arguments{
\item{x}{Covariate matrix}

\item{y}{Response vector}

\item{beta_guess1}{Initial guess of beta for m1}

\item{beta_guess0}{Initial guess of beta for m0}

\item{imp.solver}{Specifies which solver to be used. Current options
optim and cobyla (from nloptr package).}

\item{imp.kernel}{Specifies which kernel function to be used,
current options are: "EPAN", "QUARTIC", and
"GAUSSIAN".}

\item{imp.explicit_bandwidth}{Specifies if bandwidth_scale will be used as the
bandwidth or if it will be calculated as bw =
bandwidth_scale * sd(x * beta) * n^(1/3).}

\item{imp.recalc_bandwidth}{Specifies wheter the bandwidth should be
recalculated after the estimation of alpha
(cms.ps.semi).}

\item{bwc_dim_red1}{Scaling of calculated bandwidth, or if
explicit_bandwidth = TRUE used as the banddwidth.
For dimension reduction (cms.semi).}

\item{bwc_impute1}{Scaling of calculated bandwidth, or if
explicit_bandwidth = TRUE used as the banddwidth.
Recalculated if explicit_bandwidth = FALSE and
recalc_bandwidth = TRUE. For imputation.}

\item{bwc_dim_red0}{See bwc_dim_red1}

\item{bwc_impute0}{See bwc_impute1}

\item{imp.gauss_cutoff}{Cutoff value for Gaussian kernel}

\item{imp.penalty}{Penalty for the optimizer if local linear
regression fails. Added to the function value in
solver as: penalty^(n - n_before_pen), where n is
the number of llr fails.}

\item{imp.n_before_pen}{Number of probabilities outside the range (0, 1)
to accept during dimension reduction.}

\item{imp.to_extrapolate}{Specifies wheter to extrapolate or not}

\item{imp.to_truncate}{Specifies wheter to extrapolate or not}

\item{imp.extrapolation_basis}{Number of data point to base extrapolation on.}

\item{alpha_initial}{Initial guess of beta for m1}

\item{ipw.kernel}{Specifies which kernel function to be used,
current options are: "EPAN", "QUARTIC", and
"GAUSSIAN".}

\item{ipw.explicit_bandwidth}{Specifies if bandwidth_scale will be used as the
bandwidth or if it will be calculated as bw =
bandwidth_scale * sd(x * beta) * n^(1/3).}

\item{ipw.recalc_bandwidth}{Specifies wheter the bandwidth should be
recalculated after the estimation of alpha
(cms.ps.semi)}

\item{bwc_dim_red}{Scaling of calculated bandwidth, or if
explicit_bandwidth = TRUE used as the banddwidth.
For dimension reduction (cms.ps.semi).}

\item{bwc_prop_score}{Scaling of calculated bandwidth, or if
explicit_bandwidth = TRUE used as the banddwidth.
Recalculated if explicit_bandwidth = FALSE and
recalc_bandwidth = TRUE. For propensity score.}

\item{ipw.gauss_cutoff}{cutoff value for Gaussian kernel}

\item{ipw.penalty}{Penalty for the optimizer if a probability is
outside (0, 1) during dimension reduction. Added
to the function value in solver as: penalty^(n -
n_before_pen), where n is the number of
probabilities outside (0, 1).}

\item{ipw.n_before_pen}{Number of probabilities outside the range (0, 1)
to accept during dimension reduction.}

\item{n_threads}{Sets number of threads for parallel run. Set to 0
serial. If n_threads exceeds maximum number of
threads, sets n_threads to max_threads - 1. To
use max_threads, set to n_threads to max_threads
of system.}

\item{verbose}{Specifies if the program should print output
while running.}

\item{treated1}{Binary vector indicating treatment.}

\item{...}{Additional parameters passed to optim.}
}
\value{
A list containing the average treatment effect of the
        combination of observed and imputed values (ate), the average
        treatment effect based on the imputed values only (ate2), the
        imputed values for treated (m1) and untreated treated (m0), the and
        the output from optim (op).
}
\description{
Semiparametric estimation of the average treatment effect based
             on the all methods described in Ghosh, Ma, & De Luna
             (2020) and all infrences.
}
\examples{
# Using example data from package SDRcausal
library(SDRcausal)

# Import example data
x <- SDRcausal::covariates
y <- SDRcausal::outcomes
trt <- SDRcausal::treated
b1 <- SDRcausal::beta1_guess
b0 <- SDRcausal::beta0_guess
a <- SDRcausal::alpha_initial

# Perform semiparametric imputation
inf.ate <- SDRcausal::inf.ate(x, y, trt, b1, b0, alpha_initial = a)

}
\references{
Ghosh, T., Ma, Y., & De Luna, X. (2020). Sufficient dimension
reduction for feasible and robust estimation of average causal effect.
Statistica Sinica, accepted.
}
\seealso{
[stats::optim]
}
